+++ src/org/regenstrief/linkage/io/DataBaseRecordStore.java
@@ -176,7 +176,7 @@ public class DataBaseRecordStore implements RecordStore {
 	 */
 	public boolean storeRecord(Record r) {
 		try{
-			insert_stmt.setInt(1, r.getUID());
+			insert_stmt.setLong(1, r.getUID());
 			insert_count++;
 			for(int i = 0; i < insert_demographics.size(); i++){
 				String demographic = insert_demographics.get(i);
